requirements.txt
Flask==0.9
Flask-Assets==0.8
Flask-OAuth==0.12
Flask-SQLAlchemy==0.16
Flask-Script==0.5.3
MySQL-python==1.2.4
httplib2==0.8
ipython==0.13.1
requests==1.2.0
tornado==3.0
webassets==0.8
Flask-Github==0.1.3
ejson==0.1.3
markdown2==2.1.0
alembic==0.5.0
markment==0.0.1